refactor camera logic + fix pkg name
