+++ app/build.gradle
@@ -16,7 +16,7 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
-    namespace 'com.poopjournal.flashy'
+    namespace 'rocks.poopjournal.flashy'
 }
 
 dependencies {
