+++ app/src/main/java/rocks/poopjournal/flashy/GrantSystemWritePermissionActivity.java
@@ -1,4 +1,4 @@
-package com.poopjournal.flashy;
+package rocks.poopjournal.flashy;
 
 import android.app.Activity;
 import android.content.Intent;
